refactor: 리사이클러뷰 아이템 data.dto에서 domain.model로 수정
- 현재 ui에서 사용되고 있는 dto 클래스 교체

+++ app/src/main/java/com/woowa/banchan/domain/model/FoodItem.kt
@@ -1,13 +1,11 @@
 package com.woowa.banchan.domain.model
 
-import com.google.gson.annotations.SerializedName
-
 data class FoodItem(
     val description: String,
     val detailHash: String,
     val image: String,
-    val nPrice: Long?,
-    val sPrice: Long,
+    val nPrice: Int?,
+    val sPrice: Int,
     val percent: Int?,
     val title: String
 )
